Catch basic configuration errors and display to user.
E.g. If the asset builder has no "Asset Types to Create" previously this
plugin would appear to be uploading the files, but nothing would
actually get created in Matrix. Now at least the text from Matrix is
appended to the screen:

   The new asset was not created. Ensure you have entered all required
   information and try again.
    * Create type is not allowed

This could probaly do with a cleaner strategy, but as a first pass it
works.